backend/agents/sourcing_agent.py
"""Gemini-powered agent for lead sourcing from various sources"""
import re
import asyncio
from typing import Dict, List, Optional
import requests
from bs4 import BeautifulSoup


class LeadSourcingAgent:
    """Gemini-powered agent for lead sourcing from various sources"""

    def __init__(self, config: Dict):
        self.config = config
        # In production: Initialize Gemini client
        # import google.generativeai as genai
        # genai.configure(api_key=config.get('api_key'))
        # self.model = genai.GenerativeModel(config.get('model', 'gemini-1.5-pro'))

    async def scan_sources(self, sources: List[str]) -> List[Dict]:
        """Scan Google Drive, Docs, and web for potential leads"""
        leads = []

        for source in sources:
            if source.startswith('http'):
                leads.extend(await self._scan_website(source))
            elif 'drive.google.com' in source:
                leads.extend(await self._scan_drive(source))
            elif source.endswith('.pdf') or source.endswith('.docx'):
                leads.extend(await self._scan_document(source))

        return leads

    async def _scan_website(self, url: str) -> List[Dict]:
        """Scan property listing websites"""
        # Simplified example - in production use Gemini's web scraping capabilities
        try:
            # Add headers to mimic browser request
            headers = {
                'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36'
            }

            response = requests.get(url, headers=headers, timeout=10)

            if response.status_code != 200:
                print(f"  Warning: Could not access {url} (status {response.status_code})")
                return []

            soup = BeautifulSoup(response.text, 'html.parser')

            # Extract property data (example patterns)
            properties = []

            # Look for common property listing patterns
            listing_selectors = [
                'div.listing',
                'div.property',
                'article.property-card',
                'div.property-item'
            ]

            listings = []
            for selector in listing_selectors:
                found = soup.select(selector)
                if found:
                    listings.extend(found)
                    break

            # If no specific listings found, try generic approach
            if not listings:
                listings = soup.find_all('div', class_=re.compile(r'listing|property', re.I))[:5]

            for listing in listings[:10]:  # Limit to first 10 found
                address = self._extract_address(listing.get_text())
                owner = self._extract_owner(listing.get_text())

                if address:
                    properties.append({
                        'address': address,
                        'owner': owner or "Owner Name Pending",
                        'source': url,
                        'raw_text': listing.get_text()[:500]  # Truncate for efficiency
                    })

            # Fallback: Create sample data if nothing found (for testing)
            if not properties:
                print(f"  Note: No properties extracted from {url}, using sample data")
                properties = self._generate_sample_leads(url)

            return properties

        except Exception as e:
            print(f"  Error scanning website {url}: {e}")
            # Return sample data for testing purposes
            return self._generate_sample_leads(url)

    async def _scan_drive(self, drive_url: str) -> List[Dict]:
        """Scan Google Drive documents for lead information"""
        # In production: Use Google Drive API with Gemini to parse documents
        # This is a placeholder for actual implementation
        print("  Note: Google Drive scanning not yet implemented")
        return []

    async def _scan_document(self, file_path: str) -> List[Dict]:
        """Scan local documents for lead information"""
        # In production: Parse PDF/DOCX and use Gemini to extract structured data
        print(f"  Note: Document scanning not yet implemented for {file_path}")
        return []

    def _extract_address(self, text: str) -> Optional[str]:
        """Extract address using regex patterns"""
        patterns = [
            r'\d+\s+[\w\s]+(?:Street|St|Avenue|Ave|Road|Rd|Boulevard|Blvd|Lane|Ln|Drive|Dr|Court|Ct|Way|Place|Pl)',
            r'\d+\s+[\w\s]+,\s*[\w\s]+,\s*[A-Z]{2}\s*\d{5}'
        ]

        for pattern in patterns:
            match = re.search(pattern, text, re.IGNORECASE)
            if match:
                return match.group().strip()
        return None

    def _extract_owner(self, text: str) -> Optional[str]:
        """Extract owner name using simple patterns"""
        # In production, use Gemini's NER capabilities
        # This is a simplified placeholder
        owner_patterns = [
            r'Owner:\s*([A-Z][a-z]+\s+[A-Z][a-z]+)',
            r'Contact:\s*([A-Z][a-z]+\s+[A-Z][a-z]+)',
        ]

        for pattern in owner_patterns:
            match = re.search(pattern, text)
            if match:
                return match.group(1)

        return None

    def _generate_sample_leads(self, source: str) -> List[Dict]:
        """Generate sample leads for testing purposes"""
        return [
            {
                'address': '123 Main Street, Brooklyn, NY 11201',
                'owner': 'John Smith',
                'source': source,
                'raw_text': 'Sample property listing for testing'
            },
            {
                'address': '456 Oak Avenue, Queens, NY 11375',
                'owner': 'Jane Doe',
                'source': source,
                'raw_text': 'Sample property listing for testing'
            },
            {
                'address': '789 Elm Road, Manhattan, NY 10001',
                'owner': 'Robert Johnson',
                'source': source,
                'raw_text': 'Sample property listing for testing'
            }
        ]
